Logic: Running State
Also duplicated the jumping and falling states so there is a running jump and falling to maintain running speed boost. Moving running to a toggle in state.gd may be an option in the future, but getting this version causes the player to be stuck in those states, which is intended.